+++ example/random_string_example.dart
@@ -8,6 +8,5 @@ main() {
   print(randomNumeric(4)); // sequence of 4 random numbers i.e. 3259
   print(randomString(10)); // random sequence of 10 characters i.e. e~f93(4l-
   print(randomAlpha(5)); // random sequence of 5 alpha characters i.e. aRztC
-  print(randomAlphaNumeric(
-      10)); // random sequence of 10 alpha numeric i.e. aRztC1y32B
+  print(randomAlphaNumeric(10)); // random sequence of 10 alpha numeric i.e. aRztC1y32B
 }