+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Changelog
 
-## 0.0.1
+## 1.0.0
 
-- Initial version, created by Stagehand
-- Added randomBetween, randomString, randomNumeric, randomAlpha, randomAlphaNumeric, and randomMerge.
+- Added tests
+- Changed sdk environment to '>=2.1.0 <3.0.0'
 
 ## 0.0.2
 
 - Added Support for Dart 2
 - Fixed Error: The argument type 'dart.core::List<dynamic>'
+
+## 0.0.1
+
+- Initial version, created by Stagehand
+- Added randomBetween, randomString, randomNumeric, randomAlpha, randomAlphaNumeric, and randomMerge.
+